keyboard shortcuts: add jk, Enter, Esc for news item navigation
in addition to the arrow keys, this allows to use the jk keys to be
used to navigate through the news items.
These keys are widely used for navigation and allow the fingers to rest
at their natural position on the keyboard.
Using the Escape to exit the news item detail view is intuitive
and using enter to open the item as well.